db8060a: point at new prebuilt images
The previous prebuilt images were "broken" by the recent change to the Quick
Start Guide. The previous boot.img looked for the root partition in "system"
whereas the instructions were changed so that the root partition was in
"userdata".  These new prebuilt images are consistent with the Quick Start
Guide.